requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
pandas>=2.2.0
psycopg2-binary
python-dotenv
pytest
pytest-cov
pytest-mock
pytest-asyncio
azure-storage-blob
azure-storage-queue
httpx 